+++ src/PhpSpreadsheet/Worksheet/AutoFilter.php
@@ -595,7 +595,9 @@ class AutoFilter
             sort($dataValues);
         }
 
-        return array_pop(array_slice($dataValues, 0, $ruleValue));
+        $slice = array_slice($dataValues, 0, $ruleValue);
+
+        return array_pop($slice);
     }
 
     /**
