+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/DbTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class DbTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerDB
+     *
+     * @param mixed $expectedResult
+     */
+    public function testDB($expectedResult, ...$args): void
+    {
+        $result = Financial::DB(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerDB()
+    {
+        return require 'tests/data/Calculation/Financial/DB.php';
+    }
+}
