+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/DdbTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class DdbTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerDDB
+     *
+     * @param mixed $expectedResult
+     */
+    public function testDDB($expectedResult, ...$args): void
+    {
+        $result = Financial::DDB(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerDDB()
+    {
+        return require 'tests/data/Calculation/Financial/DDB.php';
+    }
+}
