+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/DiscTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class DiscTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerDISC
+     *
+     * @param mixed $expectedResult
+     */
+    public function testDISC($expectedResult, ...$args): void
+    {
+        $result = Financial::DISC(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerDISC()
+    {
+        return require 'tests/data/Calculation/Financial/DISC.php';
+    }
+}
