+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/FvTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class FvTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerFV
+     *
+     * @param mixed $expectedResult
+     */
+    public function testFV($expectedResult, ...$args): void
+    {
+        $result = Financial::FV(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerFV()
+    {
+        return require 'tests/data/Calculation/Financial/FV.php';
+    }
+}
