+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/IrrTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class IrrTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerIRR
+     *
+     * @param mixed $expectedResult
+     */
+    public function testIRR($expectedResult, ...$args): void
+    {
+        $result = Financial::IRR(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerIRR()
+    {
+        return require 'tests/data/Calculation/Financial/IRR.php';
+    }
+}
