+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/NPerTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class NPerTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerNPER
+     *
+     * @param mixed $expectedResult
+     */
+    public function testNPER($expectedResult, ...$args): void
+    {
+        $result = Financial::NPER(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerNPER()
+    {
+        return require 'tests/data/Calculation/Financial/NPER.php';
+    }
+}
