+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/NpvTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class NpvTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerNPV
+     *
+     * @param mixed $expectedResult
+     */
+    public function testNPV($expectedResult, ...$args): void
+    {
+        $result = Financial::NPV(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerNPV()
+    {
+        return require 'tests/data/Calculation/Financial/NPV.php';
+    }
+}
