+++ tests/PhpSpreadsheetTests/Calculation/Functions/Financial/RateTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Financial;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Financial;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PHPUnit\Framework\TestCase;
+
+class RateTest extends TestCase
+{
+    protected function setUp(): void
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+    }
+
+    /**
+     * @dataProvider providerRATE
+     *
+     * @param mixed $expectedResult
+     */
+    public function testRATE($expectedResult, ...$args): void
+    {
+        $result = Financial::RATE(...$args);
+        self::assertEqualsWithDelta($expectedResult, $result, 1E-8);
+    }
+
+    public function providerRATE()
+    {
+        return require 'tests/data/Calculation/Financial/RATE.php';
+    }
+}
